Code cleanup and template refactoring
