+++ templates/sentry_run_template.html
@@ -19,7 +19,6 @@
 			<div id="my_camera"></div>
 			<div id="toggle-visibility" class="text-center">
 	            <p class="text-center mx-auto">Sentry has been activated</p>
-
 	            <input id ="toggle_button" type=button class="btn btn-success btn-squared" value="Hide webcam" onClick="toggle_visibility()">
         	</div>
 	    </div>
